beginning redux use
