+++ android/app/build.gradle
@@ -133,6 +133,7 @@ android {
 }
 
 dependencies {
+    compile project(':realm')
     compile project(':react-native-vector-icons')
     compile(project(':react-native-maps')){
         exclude group: 'com.google.android.gms', module: 'play-services-base'
