+++ android/build.gradle
@@ -6,7 +6,7 @@ buildscript {
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:2.3.3'
-
+        compile project(':realm')
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
