+++ package.json
@@ -5,7 +5,7 @@
   "devDependencies": {
     "babel-preset-react-native-stage-0": "^1.0.1",
     "jest-expo": "~19.0.0",
-    "react-test-renderer": "16.0.0-alpha.12"
+    "react-test-renderer": "16.0.0-alpha.13"
   },
   "scripts": {
     "start": "react-native start",
@@ -26,8 +26,9 @@
     "lodash": "^4.17.4",
     "prettier-eslint": "^6.4.2",
     "prop-types": "^15.5.10",
-    "react": "16.0.0-alpha.12",
-    "react-native": "0.47.1",
+    "react": "16.0.0-alpha.13",
+    "react-dom": "16.0.0-alpha.13",
+    "react-native": "0.47.2",
     "react-native-action-button": "^2.7.2",
     "react-native-elements": "^0.16.0",
     "react-native-keyboard-aware-scroll-view": "^0.3.0",
@@ -36,6 +37,11 @@
     "react-native-permissions": "^1.0.0",
     "react-native-touchable-icons": "^0.1.0",
     "react-native-vector-icons": "^4.3.0",
-    "react-navigation": "^1.0.0-beta.11"
+    "react-navigation": "^1.0.0-beta.11",
+    "react-redux": "^5.0.6",
+    "realm": "^1.10.3",
+    "redux": "^3.7.2",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.2.0"
   }
 }
