+++ src/actions/index.js
@@ -1,0 +1,4 @@
+import * as AuthActions from './auth'
+
+const ActionCreators = Object.assign({}, AuthActions)
+export default ActionCreators
