+++ src/actions/types.js
@@ -1,0 +1,4 @@
+export const LOGIN = 'LOGIN'
+export const LOGOUT = 'LOGOUT'
+export const SIGNUP = 'SIGNUP'
+export const CREATE_ROOT_NAVIGATOR = 'CREATE_ROOT_NAVIGATOR'
