+++ src/reducers/index.js
@@ -1,0 +1,4 @@
+import { combineReducers } from 'redux'
+import * as AuthReducers from './auth'
+
+export default combineReducers(Object.assign({}, AuthReducers))
