--- src/renderIf.js
@@ -1,10 +1,0 @@
-const isFunction = input => typeof input === 'function'
-export default predicate => (elemOrThunk) => {
-  if (predicate) {
-    if (isFunction(elemOrThunk)) {
-      return elemOrThunk()
-    }
-    return elemOrThunk
-  }
-  return null
-}
